Re-organizing scheduler and rgt_test classes, moving sole class from fundamental_types into libraries
