--- harness/fundamental_types/__init__.py
@@ -1,5 +1,0 @@
-__all__ = [
-            "rgt_state"
-          ]
-
-version = 2.0
